+++ src/app/firebase.config.ts
@@ -28,7 +28,7 @@
  *               ".write": true,
  *               ".validate": "newData.hasChildren(['char', 'userName', 'ts', 'wpm'])
  *                             && newData.child('char').isString()
- *                             && newData.child('char').val().length <= 4
+ *                             && newData.child('char').val().length <= 5
  *                             && newData.child('userName').isString()
  *                             && newData.child('userName').val().length <= 20
  *                             && newData.child('ts').val() === now